requirements.txt
django==5.1.4
djangorestframework==3.15.2
drf-spectacular==0.28.0
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10